requirements.txt
pyyaml
garminexport[impersonate_browser]
